+++ HISTORY.md
@@ -3,4 +3,8 @@
 ### Aug 22 2022
 
 - [x] Create an SDL Window with color
-- [x] Draw Grid
+- [x] Add color buffer
+- [x] Draw pixel
+- [x] Draw grid
+- [x] Draw rectangle
+- [x] Draw flat cube